examples/requirements.txt
pysmb==1.1.27
requests_ntlm
